ChooseYourLookPage improvements
- build from OptionCard instead of radios
- Correct the name to "Choose ..."
- add localisation
- add a subtitle
