+++ packages/ubuntu_desktop_installer/lib/app.dart
@@ -7,7 +7,7 @@ import 'package:yaru/yaru.dart' as yaru;
 import 'app_theme.dart';
 import 'l10n/app_localizations.dart';
 import 'pages/allocate_disk_space_page.dart';
-import 'pages/chose_your_look_page.dart';
+import 'pages/choose_your_look_page.dart';
 import 'pages/keyboard_layout_page.dart';
 import 'pages/try_or_install_page.dart';
 import 'pages/turn_off_rst_page.dart';
@@ -44,14 +44,14 @@ class UbuntuDesktopInstallerApp extends StatelessWidget {
           const LocalizationsDelegateOc(),
         ],
         supportedLocales: AppLocalizations.supportedLocales,
-        home: ChoseYourLookPage(),
+        home: ChooseYourLookPage(),
         routes: <String, WidgetBuilder>{
           Routes.tryOrInstall: (context) => TryOrInstallPage(),
           Routes.turnOffRST: (context) => const TurnOffRSTPage(),
           Routes.keyboardLayout: (context) => KeyboardLayoutPage(),
           Routes.allocateDiskSpace: (context) => AllocateDiskSpacePage(),
           Routes.updatesOtherSoftware: UpdatesOtherSoftwarePage.create,
-          Routes.choseYourLook: (context) => ChoseYourLookPage(),
+          Routes.choseYourLook: (context) => ChooseYourLookPage(),
         },
       ),
     );
