+++ packages/ubuntu_desktop_installer/lib/pages/choose_your_look_page.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/material.dart';
+import 'package:provider/src/provider.dart';
+
+import '../app_theme.dart';
+import '../widgets.dart';
+import 'wizard_page.dart';
+
+class ChooseYourLookPage extends StatelessWidget {
+  const ChooseYourLookPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    void nextPage() {
+      // TODO: what's the next page to navigate to?
+      //Navigator.pushNamed(context, Routes.?????);
+    }
+
+    final theme = context.watch<AppTheme>();
+    return LocalizedView(
+      builder: (context, lang) => WizardPage(
+        header: Text(lang.chooseYourLookPageTitle),
+        actions: <WizardAction>[
+          WizardAction(
+            label: lang.backButtonText,
+            onActivated: Navigator.of(context).pop,
+          ),
+          WizardAction(
+            label: lang.continueButtonText,
+            enabled: true,
+            onActivated: nextPage,
+          ),
+        ],
+        title: Text(lang.chooseYourLookPageTitle),
+        contentPadding: EdgeInsets.fromLTRB(20, 50, 20, 150),
+        content: Center(
+          child: ListView(
+              shrinkWrap: true,
+              scrollDirection: Axis.horizontal,
+              children: <Widget>[
+                OptionCard(
+                  imageAsset: 'assets/Theme_thumbnails-Light.png',
+                  titleText: lang.chooseYourLookPageLightSetting,
+                  bodyText: lang.chooseYourLookPageLightBodyText,
+                  selected: Theme.of(context).brightness == Brightness.light,
+                  onSelected: () {
+                    theme.apply(Brightness.light);
+                  },
+                ),
+                OptionCard(
+                  imageAsset: 'assets/Theme_thumbnails-Dark.png',
+                  titleText: lang.chooseYourLookPageDarkSetting,
+                  bodyText: lang.chooseYourLookPageDarkBodyText,
+                  selected: Theme.of(context).brightness == Brightness.dark,
+                  onSelected: () => theme.apply(Brightness.dark),
+                )
+              ]),
+        ),
+      ),
+    );
+  }
+}
